restructure: tier D — group builtin runtime helpers under ops/
