+++ jetro-core/src/builtins/ops/collection.rs
